+++ migrations/20200919142032_drop_bugs_from_presentations.js
@@ -1,0 +1,13 @@
+exports.up = function (knex) {
+  return knex.schema.table('presentations', function (table) {
+    table.dropColumn('bugs');
+    table.dropColumn('highlights_impediments');
+  });
+};
+
+exports.down = function (knex) {
+  return knex.schema.table('presentations', function (table) {
+    table.json('highlights_impediments').notNullable();
+    table.json('bugs');
+  });
+};