+++ pages/logout.jsx
@@ -7,9 +7,8 @@ const Logout = () => {
   return null;
 };
 
-export const getServerSideProps = withSession(async function({ req, res }) {
+export const getServerSideProps = withSession(async function ({ req, res }) {
   req.session.destroy();
-
   res.writeHead(302, { Location: '/' });
   res.end();
   return {